chore: consolidate eslint configs
