+++ package.json
@@ -15,7 +15,6 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "lint": "eslint packages/",
-    "lint:boundaries": "eslint -c eslint.boundaries.config.js \"packages/**/*.ts\"",
     "fallow": "fallow",
     "fallow:health": "fallow health",
     "fallow:dead-code": "fallow dead-code",
@@ -26,7 +25,7 @@
     "baseline:api": "vitest run packages/web-backend/src",
     "baseline:smoke": "vitest run packages/web-backend/src/app.test.ts --testNamePattern=\"health endpoint|auth flow|chat history API|memory API|settings API|providers API\"",
     "verify:critical-flows": "npm run baseline:smoke && vitest run packages/web-backend/src/api/modules/tasks/route.test.ts packages/web-backend/src/cronjobs.test.ts",
-    "baseline:parity": "npm run lint:boundaries && npm run baseline:unit && npm run baseline:api && npm run verify:critical-flows",
+    "baseline:parity": "npm run lint && npm run baseline:unit && npm run baseline:api && npm run verify:critical-flows",
     "smoke:agent": "tsx packages/core/test-agent.ts",
     "clean": "rm -rf packages/*/dist packages/*/*.tsbuildinfo",
     "docs:dev": "vitepress dev docs",
